Added condition for non-owners/unrelated users
A user cannot:
- access a contract if they are not its owner or not associated with it,
- associate another user with an offer they do not own,
- edit a contract if they are not its owner or not associated with it,
- edit an offer if they are not its owner,
- edit/delete a comment if they are not its owner.

+++ ZlecajGo.Application/Reviews/Commands/DeleteReview/DeleteReviewCommandHandler.cs
@@ -32,6 +32,9 @@ public class DeleteReviewCommandHandler
         var review = await reviewRepository.GetWrittenReviewForUserAsync(reviewerId, revieweeId)
                      ?? throw new NotFoundException(nameof(Review),
                          $"of user who was reviewed with id {reviewerId}");
+
+        if (review.ReviewerId != user.Id)
+            throw new NotAllowedException();
         
         await reviewRepository.DeleteReviewAsync(review);
     }
